SecureText1/SecureText/server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
authlib==1.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
slowapi==0.1.9
apscheduler==3.10.4
email-validator==2.1.0
python-dotenv==1.0.0